requirements.txt
modal
torch>=1.7.0
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.4.0
scikit-image>=0.17.2
fastapi>=0.68.0
python-multipart>=0.0.5
